Relocated tooltip builder namespace.
