+++ src/main/kotlin/gregtechlite/gtlitecore/common/metatileentity/multiblock/mega/MultiblockEPCouplingAccelerator.kt
@@ -25,9 +25,9 @@ import gregtechlite.gtlitecore.api.capability.logic.ExtendedPowerMultiblockRecip
 import gregtechlite.gtlitecore.api.extension.EUt
 import gregtechlite.gtlitecore.api.pattern.TraceabilityPredicates.fusionCoils
 import gregtechlite.gtlitecore.api.pattern.TraceabilityPredicates.getAttributeOrDefault
-import gregtechlite.gtlitecore.api.translation.MultiblockTooltipBuilder.Companion.addTooltip
-import gregtechlite.gtlitecore.api.translation.mode.OverclockMode
-import gregtechlite.gtlitecore.api.translation.mode.UpgradeMode
+import gregtechlite.gtlitecore.api.metatileentity.multiblock.MultiblockTooltipBuilder.Companion.addTooltip
+import gregtechlite.gtlitecore.api.metatileentity.multiblock.OverclockMode
+import gregtechlite.gtlitecore.api.metatileentity.multiblock.UpgradeMode
 import gregtechlite.gtlitecore.client.renderer.texture.GTLiteOverlays
 import gregtechlite.gtlitecore.common.block.adapter.GTComputerCasing
 import gregtechlite.gtlitecore.common.block.variant.science.ScienceCasing
